+++ src/help/changelog.md
@@ -1,4 +1,4 @@
-#### March 25, 2025{section=changelog}
+#### May 6, 2025{section=changelog}
 In this release of the Broker, we:
 
-* Fully migrated from FSRS to SAM.gov to be the source of subaward data on the Data Broker and USASpending.gov. Note this may affect the data produced from File F generation.
+* Updated the Subaward section for the Data Sources help page after migrating to SAM.